feat: 트랜잭션 처리 로직을 TransactionTemplate으로 분리
